requirements.txt
blinker==1.4
click==6.6
feedparser==5.2.1
Flask==0.11.1
Flask-Humanize==0.3.0
flask-mongoengine==0.7.5
Flask-Script==2.0.5
Flask-WTF==0.12
gunicorn==19.6.0
humanize==0.5.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mongoengine==0.10.6
openlibrary==0.0.1
passlib==1.6.5
pymongo==3.2.2
python-slugify==1.2.0
requests==2.10.0
Unidecode==0.4.19
Werkzeug==0.11.10
wheel==0.24.0
WTForms==2.1
